requirements.txt
requests
pandas
numpy
folium
shapely
pyyaml
